refactor(cli): reorganize files
