+++ libNOM.cli/Executor.cs
@@ -1,5 +1,4 @@
-﻿using libNOM.cli.Args;
-using libNOM.io.Settings;
+﻿using libNOM.io.Settings;
 
 namespace libNOM.cli;
 
@@ -50,27 +49,6 @@ public partial class Executor
     {
         var msg = string.Format(message, interpolation ? "yes" : "no");
         Console.WriteLine($"{"".PadLeft(indentionLevel * INDENTION_SIZE)}{msg}");
-    }
-
-    #endregion
-
-    // //
-
-    #region Convert
-
-    [
-        ArgActionMethod,
-        ArgDescription("Convert a save file from any format to any format."),
-    ]
-    public static void Convert(ConvertArgs args)
-    {
-        if (args.Format == Enums.FormatEnum.Json)
-            io.Global.Convert.ToJson(args.Input.FullName, args.Output?.FullName, args.JsonIndented, args.JsonDeobfuscated);
-        else
-        {
-            var platform = Enum.Parse<io.Enums.PlatformEnum>(args.Format.ToString());
-            io.Global.Convert.ToSaveFile(args.Input.FullName, platform, args.Output!.FullName);
-        }
     }
 
     #endregion
